리다이렉트로 게시판 작성시 게시판 바로가고 게시판 삭제 막할수 있습니다 참여버튼시 중복 만드는중입니다

+++ ProjectWeMeet/src/main/java/com/bsh/projectwemeet/services/ArticleService.java
@@ -88,8 +88,12 @@ public class ArticleService {
 
     public boolean deleteByIndex(int index,ArticleEntity article,HttpSession session){
 
+        article = this.articleMapper.selectArticleByIndex(index);
         UserEntity user = (UserEntity) session.getAttribute("user");
 
+        if(!Objects.equals(article.getEmail(), user.getEmail())){
+            return false;
+        }
 
         return this.articleMapper.deleteByArticle(index) > 0;
     }
